surya: imported device specific settings
there are a few system settingd that are specific to the POCO X3.

This component adds settings to control the following capabilities:
    - Dirac sound enhancer (Stock POCO X3 comes with Dirac)
    - Display refresh rate (60/90/120Hz) setting
    - Powersave refresh rate selection
    - display doze capability
    - speaker cleanup sound
All these capabilities are available in the stock ROM.

These parts are the result of a lot of various commits scattered around
many threes and extensive development by:
    Kuba Wojciechowski <nullbytepl@gmail.com>
    Akhil Narang <me@akhilnarang.dev>
    Adithya R <gh0strider.2k18.reborn@gmail.com>
    Bruno Martins <bgcngm@gmail.com>
    HarukeyUA <inazarrusnak@gmail.com>
    arian <arian.kulmer@web.de>
    Henrique Silva <jhenrique09.mcz@hotmail.com>
    AshutoshSundresh <ashutoshsundresh@gmail.com>
    Jyotiraditya <jyotiraditya@aospa.co>
    xyyx <xyyx@mail.ru>
    TheScarastic <warabhishek@gmail.com>
    Pranav Vashi <neobuddy89@gmail.com>

And, in origin, by:
    Peter Cai <peter@typeblog.net>

The work from the above authors has been cleared of the "thermal profile
switching" settings which are not implemented in this ROM, due to being
dependent on many stock prebuilt binaries which are not included.

Change-Id: Ifdd5dbd287dd2ef3efcf4bb37d2181e23ee5c347