surya: enable smooth display
Device is capable of setting screen refresh rate up to 120Hz.
Enable setting "minimum" refresh rate and "smooth display" the AOSP way.
Default refresh is 60Hz with peak refresh set to 120Hz. This is the only
way to ensure that 120Hz can be disabled by the user. This is a
limitation of the way it's setup in the upstream code.